API

Hierarchical Logging

HierarchicalLogging.HierarchicalLogger — Type

HierarchicalLogger{T<:AbstractLogger} <: AbstractLogger

An AbstractLogger consisting of one or more child loggers of type T with associated names of the form

    key₁.key₂.key₃. ... .keyₙ

Given H::HierarchicalLogger, the parent of logger L with label key₁.key₂. ... .keyⱼ is the logger P ∈ H sharing the longest prefix with L, i.e., the label for P is key₁.key₂. ... .keyᵢ for i < j maximal in H.

This collection always contains a logger at the root with an empty label.

HierarchicalLogging.HierarchicalLogger — Method

HierarchicalLogger(root_logger; [propagate] [, delimiter] [, min_logging_level])

Constructs a HierarchicalLogger with root_logger at its root.

Arguments

root_logger::AbstractLogger: Logger associated with the root node

propagate::PropagateToChildren.Mode=PropagateToChildren.None: Determines how messages send to a node are propagated to its children

delimiter::Char='.': Parent-child delimiter in the string representation of a node

min_logging_level::LogLevel=All: Sets the initial minimum logging level for all loggers

HierarchicalLogging.MutableLogLevel — Type

MutableLogLevel

A type representing a LogLevel that can be modified in-place

HierarchicalLogging.MutableLogLevelLogger — Type

MutableLogLevel{T<:AbstractLogger} <: AbstractLogger

A type associating a log level and name with an underlying logger::AbstractLogger

HierarchicalLogging.MutableLogLevelLogger — Method

MutableLogLevelLogger(base_logger::AbstractLogger,        initial_log_level::LogLevel, name::String)

Returns a MutableLogLevelLogger with initial_log_level and name

HierarchicalLogging.closest_registered_logger — Function

closest_registered_logger(h::HierarchicalLogger{T}, key) -> T

If haskey(h, key), returns the logger associated with key.

Otherwise, returns the logger L ∈ h whose asssociated key K satisfies startswith(key, K) and the key of L is of maximal length among all such loggers in h.

HierarchicalLogging.insert_logger! — Function

insert_logger!(loggers::HierarchicalLogger, key, logger, [level=min_enabled_level(logger)])

Adds logger with associated key to loggers, if it does not exist.

Throws an ArgumentException if there is an existing logger in loggers with key

HierarchicalLogging.min_enabled_level! — Method

min_enabled_level!(h::HierarchicalLogger, level; [force], [recurse]) -> LogLevel

Sets the logging level of the root logger in `h` to `level`. Any messages the root logger receives below `level` will be discarded and not logged.

Returns the minimum `LogLevel` over all loggers in `h`.

HierarchicalLogging.min_enabled_level! — Method

min_enabled_level!(h::HierarchicalLogger, key, level; [force=true], [recurse=true]) -> LogLevel

Recursively sets the level of logger L associated with key in h.

If haskey(h, key), sets the logging level of L (and, if recurse == true, the level of all of its children) to level. If the current logging level of L is greater than level, this method will not change the existing level if force == false.

If haskey(h, key) == false, a logger is added to key (with base logger underlying_logger(closest_logger(h, key)))

Returns the new LogLevel associated with key

HierarchicalLogging.min_enabled_level! — Method

min_enabled_level!(logger::MutableLogLevelLogger, level) -> LogLevel

Sets minimum enabled level for logger to level. e.g., the low level below or equal to which all messages are filtered.

HierarchicalLogging.set_logger! — Method

set_logger!(loggers, key, logger)

Sets the logger associated with key in loggers to logger.

If there is an existing value associated with key in loggers, it is overwritten. The new level of key is set to the maximum of the previous log level associated to key and the current level for logger.

HierarchicalLogging.underlying_logger — Method

underlying_logger(logger::MutableLogLevel{T}) -> T

Returns the underlying logging object associated with logger

HierarchicalLogging.ROOT_LOGGER_KEY — Constant

ROOT_LOGGER_KEY

Logger key to which all other loggers are children